oss-news-agent/requirements.txt
gradio
openai
python-dotenv
trafilatura
langfuse